Make jitterMaxPercent a Double to match schema

+++ web-telemetry/web-telemetry-impl/src/main/java/com/duckduckgo/webtelemetry/impl/WebTelemetryConfig.kt
@@ -85,7 +85,7 @@ data class PixelTriggerConfig(
  */
 data class PixelPeriodConfig(
     val days: Int,
-    val jitterMaxPercent: Int,
+    val jitterMaxPercent: Double,
 ) {
     val periodSeconds: Long get() = days.toLong() * 86400
 }
